document requirement for gsed on mac (#64)
* document requirement for gsed on mac

* adjust for sed command on mac

* PR feedback: make script self-document gsed requirement on mac

+++ build_diffoscope.sh
@@ -7,7 +7,14 @@ diffoscope_file=$(dirname ${compare})/$(basename ${compare} .buildinfo.compare).
 
 echo -e "saving build diffoscope file to \033[1m${diffoscope_file}\033[0m"
 
-grep '# diffoscope ' ${compare} | sed -e 's/# diffoscope //' | ( while read -r line
+set sed="sed"
+if [ "$(uname -s)" ==  "Darwin" ]
+then
+  command -v gsed >/dev/null 2>&1 || { echo "require GNU sed: brew install gsed  Aborting."; exit 1; }
+  sed="gsed"
+fi
+
+grep '# diffoscope ' ${compare} | ${sed} -e 's/# diffoscope //' | ( while read -r line
 do
   echo -e "\033[1m$line\033[0m"
   docker run --rm -t -w /mnt -v $(pwd)/$(dirname ${compare})/${builddir}:/mnt:ro registry.salsa.debian.org/reproducible-builds/diffoscope --no-progress $line
@@ -15,7 +22,7 @@ do
 done ) | tee ${diffoscope_file}
 
 # remove ansi escape codes from file
-sed -i 's/\x1b\[[0-9;]*m//g' ${diffoscope_file}
+${sed} -i 's/\x1b\[[0-9;]*m//g' ${diffoscope_file}
 
 echo -e "build diffoscope file saved to \033[1m${diffoscope_file}\033[0m"
 du -h ${diffoscope_file}
